requirements.txt
requests
pillow
fastapi
uvicorn
transformers
torch
datasets
jupyter
ipykernel
accelerate
python-multipart